Abacus KO: improve product page layout
Move `Layout` component outside of the query renderer to prevent jitters.
