+++ src/abacus-kochka/translations/source_strings.json
@@ -523,10 +523,10 @@
     "M7rR/kOLux0twb9muLT+Nw==": "Online shop"
    },
    "filepath": "src/shop/ProductPageLayout.js",
-   "line_beg": 49,
-   "col_beg": 19,
-   "line_end": 49,
-   "col_end": 64,
+   "line_beg": 13,
+   "col_beg": 13,
+   "line_end": 13,
+   "col_end": 58,
    "desc": "shop page title",
    "project": "",
    "type": "text",
@@ -537,10 +537,10 @@
     "bQxirHhLBpNG49T54JXMeg==": "Support our cats by buying some of our products"
    },
    "filepath": "src/shop/ProductPageLayout.js",
-   "line_beg": 51,
-   "col_beg": 14,
-   "line_end": 51,
-   "col_end": 98,
+   "line_beg": 15,
+   "col_beg": 8,
+   "line_end": 15,
+   "col_end": 92,
    "desc": "shop page subtitle",
    "project": "",
    "type": "text",
@@ -550,11 +550,11 @@
    "hashToText": {
     "bU6rVH2V6ifH2RNffzFZ8g==": "All our products are currently available only in person in our café. We are working on making them available online as well."
    },
-   "filepath": "src/shop/ProductPageLayout.js",
-   "line_beg": 81,
-   "col_beg": 18,
-   "line_end": 84,
-   "col_end": 24,
+   "filepath": "src/shop/ProductPageLayoutContent.js",
+   "line_beg": 73,
+   "col_beg": 16,
+   "line_end": 76,
+   "col_end": 22,
    "desc": "not about all our products being available only in person",
    "project": "",
    "type": "text",
