+++ src/abacus-kochka/translations/src_manifest.json
@@ -1,1 +1,1 @@
-{"translations/enum_manifest.json":["src/Homepage.js","src/HomepageFooter.js","src/HomepageNavigation.js","src/LayoutFooter.js","src/LayoutNavigation.js","src/menu/MenuLoader.js","src/menu/MenuSectionCoffee.js","src/menu/MenuSectionKochkadaSavory.js","src/menu/MenuSectionKochkadaSweet.js","src/menu/MenuSectionMilkshake.js","src/menu/MenuSectionSpecialities.js","src/menu/MenuSectionTea.js","src/Rules.js","src/shop/ProductPageLayout.js","src/shop/ShopLayout.js","src/shop/ShopLayoutContent.js","src/SocialMediaIcons.js","pages/_app.js","pages/404.js","pages/adopt.js","pages/cart.js","pages/legal/privacy.js","pages/legal/shipping.js","pages/legal/terms.js","pages/menu.js","pages/rules.js","pages/shop/checkout/cancel.js","pages/shop/checkout/success.js"]}
+{"translations/enum_manifest.json":["src/Homepage.js","src/HomepageFooter.js","src/HomepageNavigation.js","src/LayoutFooter.js","src/LayoutNavigation.js","src/menu/MenuLoader.js","src/menu/MenuSectionCoffee.js","src/menu/MenuSectionKochkadaSavory.js","src/menu/MenuSectionKochkadaSweet.js","src/menu/MenuSectionMilkshake.js","src/menu/MenuSectionSpecialities.js","src/menu/MenuSectionTea.js","src/Rules.js","src/shop/ProductPageLayout.js","src/shop/ProductPageLayoutContent.js","src/shop/ShopLayout.js","src/shop/ShopLayoutContent.js","src/SocialMediaIcons.js","pages/_app.js","pages/404.js","pages/adopt.js","pages/cart.js","pages/legal/privacy.js","pages/legal/shipping.js","pages/legal/terms.js","pages/menu.js","pages/rules.js","pages/shop/checkout/cancel.js","pages/shop/checkout/success.js"]}
